octofit-tracker/backend/requirements.txt
flask
flask-restful
flask-sqlalchemy
pymongo
Django==4.1
django-cors-headers